refactor: reduce warmup time and complexity
currently, when we start a SeAT container, it's done in 3 stages
1) pull & install standard packages
2) pull & install plugins
3) pull & install dev override

at every stage, we run composer update, dump-autoload and publish commands which is a non sense
indeed, user will never had time to do anything with its container until last stage

as a result, this is merging all stages in a single one in order to reduce call against composer
1) ask for plugins
2) add dev override
3) regenerate class loader
4) publish vendors
5) update cache
6) run migrations (web)

by the meantime, this refactor is also attempt to improve files permissions update by reducing appliance time and complexity.

until now, we were searching for all files except packages directory which is outside container and not writeable - then applying permission file after file.
it appears files which must be updated are only located into storage path
as a result, we update permissions globally only on that very path